fix(nemotron): align Hindi model in BF16
Use BF16 for both the TensorRT bundle and Hugging Face reference so generation validation compares the model in its supported reduced precision without falling back to FP32.

+++ tests/e2e/models/nemotron/test_nemotron_manifest_contract.py
@@ -13,5 +13,5 @@ def test_nemotron_hindi_reference_matches_bundle_precision() -> None:
     manifest_path = Path(__file__).with_name("manifests") / "nemotron-hindi-4b.json"
     manifest = json.loads(manifest_path.read_text(encoding="utf-8"))
 
-    assert manifest["precision"] == "fp32"
+    assert manifest["precision"] == "bf16"
     assert manifest["testcases"][0]["reference_precision"] == manifest["precision"]
